Directury restructure, documented code and fixed UI and responsive bugs
