+++ src/components/homepage_components/MainContent.js
@@ -1,5 +1,4 @@
 import React, { useState, useEffect } from "react";
-import ReactDOM from "react-dom";
 import { Link } from "react-router-dom";
 
 const MainContent = (fromSiblings) => {
@@ -8,26 +7,20 @@ const MainContent = (fromSiblings) => {
   const [prevLink, setPrevLink] = React.useState("");
   const [loadNo, setLoadNo] = React.useState(1);
   const [loader, setLoader] = React.useState(1);
+  const [isFetching, setIsFetching] = useState(false);
 
+  // To set selected topics
   const onClickTopic = (e) => {
     fromSiblings.setPassedTopic(e);
   };
-
-
-  const [isFetching, setIsFetching] = useState(false);
+  
+  // To check scroll
   useEffect(() => {
     window.addEventListener('scroll', handleScroll);
     return () => window.removeEventListener('scroll', handleScroll);
   }, []);
 
-  useEffect(() => {
-    console.log("ujjwal");
-    if (!isFetching) return;
-    setLoader("loading");
-    fetchMoreListItems();
-    setLoader("hideLoading");
-  }, [isFetching]);
-
+  // handle stroll to bottom
   function handleScroll() {
     var x = document.documentElement.scrollTop + 1;
     if (window.innerHeight + x <= document.documentElement.offsetHeight || isFetching) return;
@@ -35,70 +28,75 @@ const MainContent = (fromSiblings) => {
     setIsFetching(true);
   }
 
+  // To fetch on scroll to bottom
+  useEffect(() => {
+    if (!isFetching) return;
+    setLoader("loading");
+    fetchMoreListItems();
+    setLoader("hideLoading");
+  }, [isFetching]);
+
+  // To fetch called from useEffect
   function fetchMoreListItems() {
     setTimeout(() => {
-          console.log("fetching")
-          try {
-                let link = prevLink + "?pageNo=" + loadNo,
-                  cond = false;
-                console.log(link);
-                fetch(link)
-                  .then((results) => results.json())
-                  .then((data) => {
-                    for (var i = 0; i < data.length; i++) {
-                      var companyLink = data[i].link.substring(8, data[i].link.length);
-                      var imgLink = "";
-                      for (let j = 0; j < companyLink.length; j++) {
-                        if (companyLink[j] === "/") break;
-                        imgLink += companyLink[j];
-                      }
-                      if (imgLink === "tech.ebayinc.comhttps:")
-                        imgLink = "tech.ebayinc.com";
-                      imgLink = "//logo.clearbit.com/" + imgLink + "?size=180";
-                      // console.log(imgLink);
-                      var x = {
-                        id: data[i].id,
-                        author: data[i].author,
-                        company: data[i].company,
-                        link: data[i].link,
-                        date: data[i].date,
-                        blogAbstract: data[i].blogAbstract,
-                        title: data[i].title,
-                        uuid: data[i].uuid,
-                        likes: data[i].likes,
-                        views: data[i].views,
-                        comments: data[i].comments,
-                        keywords: data[i].keywords,
-                        imgLink: imgLink,
-                        commentLink: "comment/" + data[i].id,
-                      };
-                      if (blog.indexOf(x) === -1) {
-                          blog.push(x);
-                      }
-                    }
-                  }); // fetching latest blog
-              } catch (err) {
-                alert(err); // Failed to fetch
+      try {
+        let link = prevLink + "?pageNo=" + loadNo;
+        console.log(link);
+        fetch(link)
+          .then((results) => results.json())
+            .then((data) => {
+              for (var i = 0; i < data.length; i++) {
+                var companyLink = data[i].link.substring(8, data[i].link.length);
+                var imgLink = "";
+                for (let j = 0; j < companyLink.length; j++) {
+                  if (companyLink[j] === "/") break;
+                  imgLink += companyLink[j];
+                }
+                if (imgLink === "tech.ebayinc.comhttps:")
+                  imgLink = "tech.ebayinc.com";
+                imgLink = "//logo.clearbit.com/" + imgLink + "?size=180";
+                var x = {
+                  id: data[i].id,
+                  author: data[i].author,
+                  company: data[i].company,
+                  link: data[i].link,
+                  date: data[i].date,
+                  blogAbstract: data[i].blogAbstract,
+                  title: data[i].title,
+                  uuid: data[i].uuid,
+                  likes: data[i].likes,
+                  views: data[i].views,
+                  comments: data[i].comments,
+                  keywords: data[i].keywords,
+                  imgLink: imgLink,
+                  commentLink: "comment/" + data[i].id,
+                };
+                if (blog.indexOf(x) === -1) {
+                  blog.push(x);
+                }
               }
-      
+          }); // fetching latest blog
+      } catch (err) {
+        alert(err); // Failed to fetch
+      }
+
       setIsFetching(false);
       setLoadNo(loadNo + 1);
+
     }, 2000);
   }
 
 
-
+  // to call When the states change and initial call
   React.useEffect(() => {
     try {
-      let link,
-        cond = false;
+      let link, cond = false;
+      // Check for passed parameter to search
       if (
         fromSiblings.passedTopics === "" &&
         fromSiblings.passedCompany === ""
       ) {
-        link =
-          "http://localhost:8080/home/" +
-          fromSiblings.selevtedNav 
+        link = "http://localhost:8080/home/" + fromSiblings.selevtedNav 
           if(prevLink != link){
             blog.length = 0;
             setLoadNo(1);
@@ -107,9 +105,7 @@ const MainContent = (fromSiblings) => {
         fromSiblings.passedCompany !== "" &&
         fromSiblings.passedTopics === ""
       ) {
-        link =
-          "http://localhost:8080/search/company/" +
-          fromSiblings.passedCompany
+        link = "http://localhost:8080/search/company/" + fromSiblings.passedCompany
           if(prevLink != link){
             blog.length = 0;
             setLoadNo(1);
@@ -118,23 +114,20 @@ const MainContent = (fromSiblings) => {
         fromSiblings.passedCompany === "" &&
         fromSiblings.passedTopics !== ""
       ) {
-        link =
-          "http://localhost:8080/search/keyword/" +
-          fromSiblings.passedTopics 
+        link = "http://localhost:8080/search/keyword/" + fromSiblings.passedTopics 
           if(prevLink != link){
             blog.length = 0;
             setLoadNo(1);
           }
       } else {
-        link =
-          "http://localhost:8080/search/keyword/" +
-          fromSiblings.passedTopics 
+        link = "http://localhost:8080/search/keyword/" + fromSiblings.passedTopics 
         cond = true;
         if(prevLink != link){
           blog.length = 0;
           setLoadNo(1);
         }
       }
+
       setPrevLink(link);
       fetch(link)
         .then((results) => results.json())
@@ -183,6 +176,8 @@ const MainContent = (fromSiblings) => {
       {blog.map((e) => (
         <div className="container">
           <div className="flex_box">
+            
+            {/* Company logo */}
             <div className="blankContainer ">
               <img
                 src={e.imgLink}
@@ -192,6 +187,7 @@ const MainContent = (fromSiblings) => {
               />
             </div>
 
+            {/* abstract, title, date, author */}
             <div className="contentContainer">
               <div className="content-heading">
                 <Link
@@ -224,6 +220,7 @@ const MainContent = (fromSiblings) => {
               </div>
             </div>
 
+            {/* Likes, comment type */}
             <div className="statsContainer">
               <div>
                 <span>
@@ -231,9 +228,7 @@ const MainContent = (fromSiblings) => {
                 </span>{" "}
                 {e.likes}
               </div>
-
               <br></br>
-
               <div>
                 <span>
                   <i class="fa fa-comment-o" aria-hidden="true"></i>{" "}
@@ -251,6 +246,8 @@ const MainContent = (fromSiblings) => {
           </div>
         </div>
       ))}
+
+      {/* Loader */}
       <div className="loading">
         <div className="loader-ellips">
             <span className="loader-ellips__dot"></span>
