+++ src/components/homepage_components/company_homepage.js
@@ -2,26 +2,30 @@ import React, { useState, useEffect } from "react";
 import ReactDOM from "react-dom";
 import "font-awesome/css/font-awesome.min.css";
 
-let tempSelectedTopics = [];
-
-const RightTags = ({ setPassedTopic }, passedTopics) => {
-  const [topic, setTopic] = React.useState([]);
-  const [topics, setTopics] = React.useState([]);
+const RightTags = ({ setPassedCompany }) => {
+  const [company, setCompany] = React.useState([]);
+  const [companies, setCompanies] = React.useState([]);
   const [search, setSearch] = React.useState("");
-  const [selectedTopics, setSelectedTopics] = React.useState(null);
-  const [topicVisible, setTopicVisible] = React.useState(5);
+  const [selectedCompanies, setSelectedCompanies] = React.useState();
+  const [companyVisible, setCompanyVisible] = React.useState(10);
 
+  // show more button
+  const onClickShowMore = () => {
+    setCompanyVisible(companyVisible + 10);
+  };
+
+  // component for selected company tag
   const SelectedTagDisplay = () => {
-    if (selectedTopics === null || selectedTopics === "") return null;
+    if (selectedCompanies == null || selectedCompanies === "") return null;
     else
       return (
         <div
           style={{ background: `#6C63FF`, color: `white`, cursor: `pointer` }}
         >
-          {selectedTopics}{" "}
+          {selectedCompanies}{" "}
           <span
             style={{ padding: `5px` }}
-            onClick={() => onClickDeselct({ selectedTopics })}
+            onClick={() => onClickDeselct(selectedCompanies)}
           >
             &#10007;
           </span>
@@ -29,62 +33,66 @@ const RightTags = ({ setPassedTopic }, passedTopics) => {
       );
   };
 
-  const onClickShowMore = () => {
-    setTopicVisible(topicVisible + 10);
-  };
-
+  // for handling search input
   const onchange = (e) => {
     setSearch(e.target.value);
-    // console.log(e.target.value);
   };
 
-  const filteredTopics = topics.filter((x) => {
-    return x.keyword.toLowerCase().indexOf(search.toLowerCase()) !== -1;
+  // filtered topic list
+  const filteredTopics = companies.filter((x) => {
+    return x.company.toLowerCase().indexOf(search.toLowerCase()) !== -1;
   });
 
+  // on selection of company
   const onClickTopic = (e) => {
-    setSelectedTopics(e);
-    setPassedTopic(e);
+    setSelectedCompanies(e);
+    setPassedCompany(e);
   };
 
+  // on deselecting companies
   const onClickDeselct = (e) => {
-    setSelectedTopics("");
-    setPassedTopic("");
+    setSelectedCompanies("");
+    setPassedCompany("");
   };
 
+  // fetching list of companies
   React.useEffect(() => {
     try {
-      fetch("http://localhost:8080/home/keywordsCount")
+      fetch("http://localhost:8080/home/blogsCount")
         .then((results) => results.json())
         .then((data) => {
           for (var i = 0; i < data.length; i++) {
             var x = {
-              keyword: data[i].keyword,
-              frequency: data[i].frequency,
+              company: data[i].company,
+              count: data[i].count,
             };
-            if (topic.indexOf(x) === -1) {
-              topic.push(x);
+            if (
+              company.indexOf(x) === -1 &&
+              x.company != null &&
+              x.count != null
+            ) {
+              company.push(x);
             }
           }
-          setTopics(topic);
+          setCompanies(company);
         }); // fetching latest blogs
     } catch (err) {
       alert(err); // Failed to fetch
     }
-  }, [topic]);
+  }, [company]);
 
   return (
     <div>
-      <div className="tagsContainerTopics">
+      <div className="tagsContainerCompanies">
         <div className="tags-heading">
-          <span class="text-300 text-lg tags-title">
-            <i class="fa fa-bookmark-o" aria-hidden="true"></i>&nbsp; Topics
+          <span class=" tags-title">
+            <i class="fa fa-building-o" aria-hidden="true"></i>&nbsp; Companies
           </span>
 
+          {/* selected tags */}
           <div className="tags-values">
             <SelectedTagDisplay />
           </div>
-
           <div className="search-tags">
             <input
               class="search-tags-input"
@@ -95,17 +103,21 @@ const RightTags = ({ setPassedTopic }, passedTopics) => {
             />
           </div>
         </div>
-
+        
+        {/* Tags unselected */}
         <div className="tags-values">
-          {filteredTopics.slice(0, topicVisible).map((e) => (
+          {filteredTopics.slice(0, companyVisible).map((e) => (
             <div
               className="tags-Values-style"
-              onClick={() => onClickTopic(e.keyword)}
+              onClick={() => onClickTopic(e.company)}
             >
-              {e.keyword} <span className="tags-counts">{e.frequency}</span>
+              {e.company}
+              <span className="tags-counts"> {e.count}</span>
             </div>
           ))}
         </div>
+
+        {/* show more button */}
         <div className="showMoreButton" onClick={() => onClickShowMore()}>
           <span class="text">
             Show More &nbsp;
