+++ src/components/homepage_components/navbar_homepage.js
@@ -1,12 +1,11 @@
-import { check } from "prettier";
-import React, { useState, useEffect } from "react";
-import ReactDOM from "react-dom";
+import React from "react";
 
 const Header = ({ setSelectedNav }) => {
   const [latest, setLatest] = React.useState("fa fa-check");
   const [liked, setLiked] = React.useState();
   const [viewed, setViewed] = React.useState();
 
+  // selction of nav button
   const onClickNav = (e) => {
     setSelectedNav(e);
     if (e === "latest") {
