+++ src/pages/login.js
@@ -1,13 +1,12 @@
-import "./App.css";
 import React, { useState, useEffect } from "react";
-import Header from "./components/header";
-import SignupForm from "./components/signupform";
+import Header from "../components/header";
+import LoginForm from "../components/loginForm";
 
 function App() {
   return (
     <div className="app">
       <Header />
-      <SignupForm />
+      <LoginForm />
     </div>
   );
 }
