+++ src/pages/signup.js
@@ -1,13 +1,14 @@
-import "./App.css";
 import React, { useState, useEffect } from "react";
-import Header from "./components/header";
-import LoginForm from "./components/loginForm";
+
+// components 
+import Header from "../components/header";
+import SignupForm from "../components/signupform";
 
 function App() {
   return (
     <div className="app">
       <Header />
-      <LoginForm />
+      <SignupForm />
     </div>
   );
 }
